requirements.txt
SQLAlchemy==1.3.0
SQLAlchemy-Utils==0.34.1
Flask-SQLAlchemy==2.2
Flask-Migrate==2.0.4
Flask-Login==0.4.0
requests-oauthlib==0.8.0
psycopg2==2.7.1
